バックエンドのソースコードにデータ型を付与
