+++ backend/types/express/index.d.ts
@@ -1,0 +1,7 @@
+import { Request } from "express";
+
+declare module "express" {
+  export interface Request {
+    userId?: number;
+  }
+}
