+++ backend/utils/getRemainTimeForSecond.ts
@@ -1,25 +1,17 @@
-const router = require("express").Router();
-const { PrismaClient } = require("@prisma/client");
+import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
 
-const dateToSeconds = (date) => {
+const dateToSeconds = (date: Date) => {
   return Math.floor(date.getTime() / 1000);
 };
 
-const ageToSeconds = (age) => {
+const ageToSeconds = (age: number) => {
   return age * 365 * 24 * 60 * 60;
 };
 
-router.get("/lifespan", async (req, res) => {
-  const { year, sex } = req.query;
-
-  if (!year || !sex) {
-    return res.status(400).json({ error: "性別と生年月日を選択してください" });
-  }
-
+const getRemainTimeForSeconds = async (sex: string, birthDate: Date) => {
   // 生年月日をDate型で取得→秒に変換
-  const birthDate = new Date(year);
   const birthDateInSeconds = dateToSeconds(birthDate);
 
   // 現在年月日を秒で取得
@@ -33,13 +25,18 @@ router.get("/lifespan", async (req, res) => {
   const lifespan = await prisma.averageLife.findUnique({
     where: { sex_year: { sex, year: String(currentYear - 1) } },
   });
+
+  if (!lifespan) {
+    // 平均寿命を取得できなかった場合、0を返却
+    // 明らかにエラーと判別できるため
+    return 0;
+  }
+
   const averageLifespan = lifespan.age;
   const averageLifespanInSeconds = ageToSeconds(averageLifespan);
 
-  // 余命を秒で算出
-  const remainingTimeInSeconds = averageLifespanInSeconds - ageInSeconds;
+  // 余命(秒)を返却
+  return averageLifespanInSeconds - ageInSeconds;
+};
 
-  return res.json({ remainTime: remainingTimeInSeconds });
-});
-
-module.exports = router;
+export default getRemainTimeForSeconds;
